move dataset db methods to their own file
